Move and add finalized state methods
